Refactors linear function synthesis tests
